% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_docker.R
\name{build_deploy}
\alias{build_deploy}
\title{Build Package and Copy renv.lock}
\usage{
build_deploy(.deploy_path = "deploy", .lock_file = "renv.lock")
}
\arguments{
\item{.deploy_path}{A character string of the deploy path}

\item{.lock_file}{A characterstring of the lock file}
}
\description{
Build the package into a tar-ball and copy renv.lock into the deploy-folder.
}
