% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_db.R
\name{db_overwrite_itemdata}
\alias{db_overwrite_itemdata}
\title{Overwrite Item Database}
\usage{
db_overwrite_itemdata(
  .file_data,
  .drv = RSQLite::SQLite(),
  .dbname = "db_item.sqlite"
)
}
\arguments{
\item{.file_data}{Absolute file path to the .csv file containing the item data}

\item{.drv}{Database driver}

\item{.dbname}{Relative path to the database}
}
\description{
Overwrite Item Database
}
