% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_filter_handler.R
\name{FilterHandler}
\alias{FilterHandler}
\title{R6 Class Item Filtering Handler}
\description{
This class handles the filtering of items based on the user's selection.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FilterHandler-new}{\code{FilterHandler$new()}}
\item \href{#method-FilterHandler-observeFilters}{\code{FilterHandler$observeFilters()}}
\item \href{#method-FilterHandler-observeSubmit}{\code{FilterHandler$observeSubmit()}}
\item \href{#method-FilterHandler-clone}{\code{FilterHandler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterHandler-new"></a>}}
\if{latex}{\out{\hypertarget{method-FilterHandler-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterHandler$new(data, filter, topics, unsolved, msg, indices, current_index)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterHandler-observeFilters"></a>}}
\if{latex}{\out{\hypertarget{method-FilterHandler-observeFilters}{}}}
\subsection{Method \code{observeFilters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterHandler$observeFilters(picker, filter_unsolved, credentials)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterHandler-observeSubmit"></a>}}
\if{latex}{\out{\hypertarget{method-FilterHandler-observeSubmit}{}}}
\subsection{Method \code{observeSubmit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterHandler$observeSubmit(submit_btn, credentials, session)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterHandler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FilterHandler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterHandler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
