% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_check_item.R
\name{mod_check_item_server}
\alias{mod_check_item_server}
\title{check_item Server Functions}
\usage{
mod_check_item_server(
  id,
  data_item,
  cur_item_id,
  cur_answer_txt,
  cur_answer_id,
  submit_btn_value
)
}
\arguments{
\item{id}{A character string of the id of the module}

\item{data_item}{A dataframe that contains the item data from the item database}

\item{cur_item_id}{test}

\item{cur_answer_txt}{test}

\item{cur_answer_id}{test}
}
\value{
Test
}
\description{
The server part of the check_item module that provides a interface with a two action buttons to check the item, receive feedback and proceed to the next item.
}
