% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_check_item.R
\name{mod_check_item_ui}
\alias{mod_check_item_ui}
\title{check_item UI Function}
\usage{
mod_check_item_ui(id)
}
\arguments{
\item{id}{A character string of the id of the module}
}
\value{
A UI definition
}
\description{
The UI part of the check_item module that provides a interface with a two action buttons to check the item, receive feedback and proceed to the next item.
}
