% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_display_item.R
\name{mod_display_item_ui}
\alias{mod_display_item_ui}
\title{display_item UI function}
\usage{
mod_display_item_ui(id)
}
\arguments{
\item{id}{A character string of the id of the module}
}
\value{
A UI definition
}
\description{
The UI part of the display_item module that creates the stimulus and MC-questionnaire interface.
}
