% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_select_item.R
\name{mod_select_item_server}
\alias{mod_select_item_server}
\title{Item Selection Module Server}
\usage{
mod_select_item_server(id, data_item, credentials)
}
\arguments{
\item{id}{Module id}

\item{data_item}{Item data dataframe}
}
\description{
Item Selection Module Server
}
