% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_select_item.R
\name{mod_select_item_ui}
\alias{mod_select_item_ui}
\title{Item Selection Module UI}
\usage{
mod_select_item_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\description{
Module creates checkbox contigency table `checkbox_table` with learning_area 
and type_item as dimensions. It also allows users to specify the number of items
to train in the learning session.
}
