% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_select_item_fct_table.R
\name{observe_margins}
\alias{observe_margins}
\title{title}
\usage{
observe_margins(input, id, checkbox_data, session)
}
\arguments{
\item{id}{description}

\item{checkbox_data}{description}

\item{session}{description}
}
\value{
description
}
\description{
A short description...
}
