% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_postanalysis.R
\name{pool_connect}
\alias{pool_connect}
\title{Establish a Database Connection Pool}
\usage{
pool_connect(.return_path)
}
\arguments{
\item{.return_path}{A character vector specifying the local paths of SQLite database files.}
}
\value{
A database connection pool object.
}
\description{
This function creates a database connection pool using the RSQLite and pool package.
The pool allows efficient management of multiple database connections, enhancing performance and resource utilization.
}
\details{
The function checks for the presence of files with a ".sqlite" extension in the specified directory.
It then creates a database connection pool using the RSQLite package for each valid SQLite file found.
The resulting pool can be used to query and manipulate data from the connected databases.
}
\seealso{
\code{\link{duck_download_files}}
}
\keyword{SQLite}
\keyword{connection}
\keyword{database}
\keyword{pool}
