% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_postanalysis.R
\name{process_run}
\alias{process_run}
\title{Execute External Process for File Download}
\usage{
process_run(.x, .cmd, arg, .user, .rsa_path, .dpath, .return_path)
}
\arguments{
\item{.x}{A character specifying the filepath on the remote server to be downloaded.}

\item{.cmd}{A character specifying the path to the external command-line executable.}

\item{arg}{A character vector specifying the arguments for the external command.}

\item{.user}{A character specifying the username for the server connection.}

\item{.rsa_path}{A character specifying the path to the RSA private key for authentication.}

\item{.dpath}{A character specifying the destination path on the remote server for the downloaded file.}

\item{.return_path}{A character specifying the local destination path to save the downloaded file.}
}
\description{
This function executes an external process using the specified command and arguments for downloading files.
It is internally used by the \code{\link{duck_download_files}} function to perform the actual file download.
}
\details{
This function is responsible for initiating the download process by executing an external command (e.g., Duck).
It provides visual feedback using the \code{cli} package to inform the user about the ongoing download.
}
\examples{
\dontrun{
# Example usage:
.cmd <- "/opt/homebrew/bin/duck"
.rsa_path <- "/Users/luca/.ssh/id_rsa"
.return_path <- c("/Users/luca/Desktop/db_user.sqlite")
.user <- "beitner"
.x <- "shinyapps/shinytigeR-DB/db_user.sqlite"
.dpath <- "sftp://tiger.uni-frankfurt.de/shinyapps/shinytigeR-DB/db_user.sqlite"
arg <- c("-u", .user, "-i", .rsa_path, "-d", .dpath, .return_path, "-e", "overwrite")

process_run(.x, .cmd, arg, .user, .rsa_path, .dpath, .return_path)
}

}
\seealso{
\code{\link{duck_download_files}}
}
\keyword{download}
\keyword{execution}
\keyword{external}
\keyword{feedback}
\keyword{process}
\keyword{visual}
