% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinytigeR-package.R
\docType{package}
\name{shinytigeR-package}
\alias{shinytigeR}
\alias{shinytigeR-package}
\title{shinytigeR: Training mit individuell generierten Erfolgsrückmeldungen in R}
\description{
A Shiny web application for Goethe University psychology students to practice statistics and R, with IRT-based skill estimation and individualised feedback.
}
\author{
\strong{Maintainer}: Jan Luca Schnatz \email{janlucaschnatz@gmail.com}

Authors:
\itemize{
  \item Julia Beitner \email{Julia.Beitner@zi-mannheim.de}
}

}
\keyword{internal}
