% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_select_item_fct_table.R
\name{update_bool_table}
\alias{update_bool_table}
\title{title}
\usage{
update_bool_table(input, id, data_handler, data_item)
}
\arguments{
\item{input}{description}

\item{id}{description}

\item{data_handler}{description}

\item{data_item}{description}
}
\value{
description
}
\description{
A short description...
}
