% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinytigeR-package.R
\docType{package}
\name{shinytigeR-package}
\alias{shinytigeR}
\alias{shinytigeR-package}
\title{shinytigeR: Training with Individually Generated Feedback in R}
\description{
An interactive Shiny app providing automated feedback and skill tracking to aid students develop statistical litercy with R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jlschnatz/shinytigeR}
  \item Report bugs at \url{https://github.com/jlschnatz/shinytigeR/issues}
}

}
\author{
\strong{Maintainer}: Jan Luca Schnatz \email{schnatz@psych.uni-frankfurt.de}

Authors:
\itemize{
  \item Julia Beitner \email{beitner@psych.uni-frankfurt.de}
}

}
\keyword{internal}
